requirements.txt
streamlit
sentence-transformers
faiss-cpu
beautifulsoup4
requests
numpy
